Cleanup IGateway.

+++ gmp/evm/gateway/src/storage/Routes.sol
@@ -145,7 +145,7 @@ library RouteStore {
         return routes;
     }
 
-    function estimateGas(NetworkInfo memory route, uint16 messageSize, uint64 gasLimit)
+    function estimateGas(NetworkInfo memory route, uint256 messageSize, uint64 gasLimit)
         internal
         pure
         returns (uint256)
@@ -153,7 +153,7 @@ library RouteStore {
         // Verify if the gas limit and message size are within the limits
         require(gasLimit <= route.gasLimit, "gas limit exceeded");
         require(messageSize <= MAX_PAYLOAD_SIZE, "maximum payload size exceeded");
-        return uint256(messageSize) * route.gasCoef1 + route.gasCoef0 + gasLimit;
+        return messageSize * route.gasCoef1 + route.gasCoef0 + gasLimit;
     }
 
     function estimateCost(NetworkInfo memory route, uint256 gas) internal pure returns (uint256) {
